Create database

+++ src/net/softwarealchemist/xpcalc/domain/Foe.java
@@ -1,6 +1,6 @@
 package net.softwarealchemist.xpcalc.domain;
 
 public class Foe extends Entity {
-	public int challengeRating;
+	public double challengeRating;
 	public int quantity;
 }
